\name{account_info}
\alias{account_info}
\title{account_info}
\description{Get Microworkers account information}
\usage{
account_info(key = getOption("microworkers_key"))
}
\arguments{
    \item{key}{A character string containing a Microworkers.com API secret key.}
}
\details{Retrieve account details for the Microworkers account specified by the API key}
%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
\examples{
\dontrun{
account_info()
}
}
%\keyword{}
