\name{microworkers-package}
\alias{microworkers-package}
\docType{package}
\title{A Survey Monkey R Client}
\description{microworkers provides access to \href{https://www.microworkers.com/}{Microworkers.com}, an easy-to-use international crowdsourcing platform.}
\details{
To use Rmonkey, the user must have a Microworkers account, which automatically supplies an API access key. To use the API, however, you must contact \href{mailto:info@microworkers.com}{info@microworkers.com} to gain API access. Once everything is registered, the relevant variables can be loaded into R using \code{options}:

\code{options(microworkers_key = 'YourAPIKey')}
}
\author{
Thomas J. Leeper

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}
\references{
\href{https://microworkers.com/api.php}{API Documentation}
}
\keyword{ package }
%\seealso{}
%\examples{}
